Ask for a lang if not specified when extracting the translations

+++ i18n/extract.sh
@@ -1,4 +1,10 @@
 
+if [ -z "$1" ]
+  then
+    echo "Please set a lang"
+    exit
+fi
+
 rm -rf generated
 mkdir generated
